+++ tybotRoute/index.js
@@ -19,6 +19,7 @@ const { TiledeskIntentsMachine } = require('./models/TiledeskIntentsMachine.js')
 // const { MockActions } = require('./MockActions');
 const { MockBotsDataSource } = require('./models/MockBotsDataSource.js');
 const { TiledeskChatbotConst } = require('./models/TiledeskChatbotConst');
+const { IntentsMachineFactory } = require('./models/IntentsMachineFactory');
 
 //router.use(cors());
 router.use(bodyParser.json({limit: '50mb'}));
@@ -111,22 +112,30 @@ router.post('/ext/:botid', async (req, res) => {
   
   let intentsMachine;
   if (!staticBots) {
-    if (log) {console.log("intentsMachine to MongoDB");}
-    intentsMachine = new MongodbIntentsMachine({projectId: projectId, language: bot.language, log});
-    if (bot.intentsEngine === "tiledesk-ai") {
-      if (log) {console.log("intentsMachine to tiledesk-ai");}
-      intentsMachine = new TiledeskIntentsMachine(
-        {
-          //projectId: projectId,
-          //language: bot.language,
-          botId: botId
-          //TILEBOT_AI_ENDPOINT: process.env.TILEBOT_AI_ENDPOINT
-        });
-    }
+    intentsMachine = IntentsMachineFactory.getMachine(bot);
   }
   else {
     intentsMachine = {}
   }
+  
+  // let intentsMachine;
+  // if (!staticBots) {
+  //   if (log) {console.log("intentsMachine to MongoDB");}
+  //   intentsMachine = new MongodbIntentsMachine({projectId: projectId, language: bot.language, log});
+  //   if (bot.intentsEngine === "tiledesk-ai") {
+  //     if (log) {console.log("intentsMachine to tiledesk-ai");}
+  //     intentsMachine = new TiledeskIntentsMachine(
+  //       {
+  //         //projectId: projectId,
+  //         //language: bot.language,
+  //         botId: botId
+  //         //TILEBOT_AI_ENDPOINT: process.env.TILEBOT_AI_ENDPOINT
+  //       });
+  //   }
+  // }
+  // else {
+  //   intentsMachine = {}
+  // }
   //const intentsMachine = new TiledeskIntentsMachine({API_ENDPOINT: "https://MockIntentsMachine.tiledesk.repl.co", log: true});
   const chatbot = new TiledeskChatbot({
     botsDataSource: botsDS,